requirements.txt
cycler==0.10.0
Cython==0.29.21
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pyjnius==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
